Nepovinny ukol L6: namockovat unit testy
