+++ ToDoList/src/ToDoList.WebApi/Program.cs
@@ -1,4 +1,6 @@
+using ToDoList.Domain.Models;
 using ToDoList.Persistence;
+using ToDoList.Persistence.Repositories;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -6,6 +8,7 @@ var builder = WebApplication.CreateBuilder(args);
     //Configure DI
     builder.Services.AddControllers();
     builder.Services.AddDbContext<ToDoItemsContext>();
+    builder.Services.AddScoped<IRepository<ToDoItem>, ToDoItemsRepository>();
 }
 
 // // Add services to the container.
